Fix #37
When setting a variable I made sure to make sure that I deep copied the expression onto the environment stack so that it would not get erased whenever I freed the expression that was in memory. However I did not do the same thing when looking up a variable on the environment stack, so lookup() returned a pointer to the deep-copied variable on the environment stack, which then got erased whenever the current line in memory was done being evaluated. The clearExpression() method only clears SExpressions, so if I was only looking at the name of the variable in the environment stack, nothing was changing because only the pointer to the expression of that variable was getting freed.