Sprint five first PR (#38)
* Updated parser_clearExpression(), cons()

Updated parser_clearExpression to return a null pointer; this is required to actually reset the pointer of whatever variable was pointing to the list that was passed to clear, otherwise the original variable won't show as null and points to an invalid memory address

Updated cons() in sexpr.c to check if both car and cdr are null and to return null if so

* Implement environment management with variable storage and retrieval

* Refactor eval function to accept Environment pointer and return updated SExpression

* Move all structs and typedefs into a types.h file to simplify necessary includes (#32)

* Add, subtract, multiply, divide, modulo working in the REPL at a single level without any nesting

* Enhance lexer and parser to support additional comparison operators; add tests for mathematical operators

* Fix #37

When setting a variable I made sure to make sure that I deep copied the expression onto the environment stack so that it would not get erased whenever I freed the expression that was in memory. However I did not do the same thing when looking up a variable on the environment stack, so lookup() returned a pointer to the deep-copied variable on the environment stack, which then got erased whenever the current line in memory was done being evaluated. The clearExpression() method only clears SExpressions, so if I was only looking at the name of the variable in the environment stack, nothing was changing because only the pointer to the expression of that variable was getting freed.

* If looking up a variable name and it does not exist, print nil

* Update set(), had same issue as described in commit 512b1d5

* Updated lexer, main, error printing, cons, and eq

Updated error handling in lexer_lex() to also print the ASCII code
Updated main to exit the repl on typing "exit"
Updated cons() to return a cons pair with two nils rather a null pointer
Updated eq() to return true if passed a cons pair with 2 nils